+++ benchmark/blankenbach/case1/gnuplot.gpl
@@ -1,0 +1,15 @@
+set term png
+set size 0.8,0.7
+set xrange[0.10:]
+
+set output "rms.png"
+plot "output/statistics" using 2:12 w l, 42.864947
+!convert rms.png -trim x.png ; mv x.png rms.png
+
+
+set output "nusselt.png"
+plot "output/statistics" using 2:21 w l, 4.884409
+!convert nusselt.png -trim x.png ; mv x.png nusselt.png
+
+
+
