+++ benchmark/blankenbach/case2/ReadMe.txt
@@ -1,0 +1,11 @@
+
+
+
+case2b:
+
+reference values: 
+Nu 6.9299, RMS 171.755
+
+ours (coarse):
+Nu = 1.83350557e+01 / 2.5 = 7.33402228
+RMS 179.245606
