+++ benchmark/tangurnis/Makefile
@@ -6,10 +6,10 @@ tala_c: tala_c/vel_8.csv tala_c/vel_16.csv tala_c/vel_32.csv tala_c/vel_64.csv t
 
 
 
-%.csv: ../../aspect ba/tan.prm tala/tan.prm tala_c/tan.prm
+%.csv: code/aspect ba/tan.prm tala/tan.prm tala_c/tan.prm
 	@echo "- generating $@"
 	@cp $(*D)/tan.prm $@_temp.prm~
 	@ref=`echo $@ | tr -cd '[[:digit:]]' | awk '{print log($$1)/log(2)}'` && (echo  "subsection Mesh refinement"; echo "set Initial global refinement = $$ref ";echo "end";echo "set Output directory=$(*D)") >> $@_temp.prm~
-	@../../aspect $@_temp.prm~
+	@./code/aspect $@_temp.prm~
 	@cd $(*D); rm -f parameters* statistics *_temp.prm~
