+++ cookbooks/convection-box.prm
@@ -99,6 +99,10 @@ subsection Model settings
   set Tangential velocity boundary indicators = left, right, bottom, top
 end
 
+subsection Formulation
+  set Formulation = custom
+  set Temperature density approximation = adiabatic
+end
 
 # The following two sections describe first the
 # direction (vertical) and magnitude of gravity and the
@@ -109,7 +113,7 @@ subsection Gravity model
   set Model name = vertical
 
   subsection Vertical
-    set Magnitude = 1e14   # = Ra / Thermal expansion coefficient
+    set Magnitude = 1e4   # = Ra / Thermal expansion coefficient
   end
 end
 
@@ -122,7 +126,7 @@ subsection Material model
     set Reference specific heat       = 1
     set Reference temperature         = 0
     set Thermal conductivity          = 1
-    set Thermal expansion coefficient = 1e-10
+    set Thermal expansion coefficient = 1
     set Viscosity                     = 1
   end
 end
