+++ doc/manual/manual.bib
@@ -11093,10 +11093,24 @@ year={2012}
   Year = {2007}
 }
 
-
 @article{GHPB17,
   Author = {R. Gassm{\"o}ller and E. Heien and E. G. Puckett and W. Bangerth},
   Title = {Flexible and scalable particle-in-cell methods for massively parallel computations},
   Journal = {submitted},
   Year = {2017}
+
+@Article{KLKLZTTK10,
+  author =       {S. D. King and C. Lee and P. {van Keken} and W. Leng and
+                  S. Zhong and E. Tan and N. Tosi and M. C. Kameyama},
+  title =        {A community benchmark for 2-{D} {C}artesian compressible 
+                  convection in {E}arth's mantl},
+  journal =      {Geophys. J. Int.},
+  year =         {2010},
+  OPTkey =       {},
+  OPTvolume =    {180},
+  OPTnumber =    {},
+  OPTpages =     {73--87},
+  OPTmonth =     {},
+  OPTnote =      {},
+  OPTannote =    {}
 }
