+++ include/aspect/simulator_access.h
@@ -64,6 +64,10 @@ namespace aspect
   {
     template <int dim> class Interface;
   }
+  namespace AdiabaticConditions
+  {
+    template <int dim> class Interface;
+  }
   template <int dim> class MeltHandler;
 
   /**
