+++ source/material_model/interface.cc
@@ -89,6 +89,7 @@ namespace aspect
       return 1.0;
     }
 
+
     template <int dim>
     void
     Interface<dim>::
