+++ source/material_model/simple.cc
@@ -27,7 +27,6 @@ namespace aspect
 {
   namespace MaterialModel
   {
-
     template <int dim>
     void
     Simple<dim>::
