+++ source/postprocess/visualization/heating.cc
@@ -144,6 +144,14 @@ namespace aspect
 
         this->get_material_model().evaluate(in, out);
 
+        if (this->get_parameters().formulation_temperature == Parameters<dim>::FormulationType::adiabatic)
+          {
+            for (unsigned int q=0; q<n_quadrature_points; ++q)
+              {
+                out.densities[q] = this->get_adiabatic_conditions().density(in.position[q]);
+              }
+          }
+
         unsigned int index = 0;
         for (typename std::list<std_cxx11::shared_ptr<HeatingModel::Interface<dim> > >::const_iterator
              heating_model = heating_model_objects.begin();
