+++ source/simulator/assembly.cc
@@ -25,7 +25,7 @@
 #include <aspect/simulator_access.h>
 #include <aspect/melt.h>
 #include <aspect/free_surface.h>
-
+#include <aspect/assemblers.h>
 
 #include <deal.II/base/quadrature_lib.h>
 #include <deal.II/base/work_stream.h>
@@ -43,30 +43,6 @@
 
 namespace aspect
 {
-  namespace
-  {
-    /* These functions implement a reduced form of the code from deal.II's TriaAccessor::measure().
-     * In the 3d dG case, a call to face->measure() is not implemented for non-planar faces.
-     * Since we only care about the scaling here, it is enough to have an approximation instead.
-     * The 2d case remains unchanged.
-     */
-    double
-    approximate_face_measure(const DoFHandler<2>::face_iterator &face)
-    {
-      return (face->vertex(0)-face->vertex(1)).norm();
-    }
-
-    double
-    approximate_face_measure(const DoFHandler<3>::face_iterator &face)
-    {
-      const Tensor<1,3> v03 = face->vertex(3) - face->vertex(0);
-      const Tensor<1,3> v12 = face->vertex(2) - face->vertex(1);
-      const Tensor<1,3> twice_area = cross_product_3d(v03, v12);
-      return 0.5 * twice_area.norm();
-    }
-  }
-
-
   namespace internal
   {
     namespace Assembly
@@ -74,10 +50,6 @@ namespace aspect
       namespace Scratch
       {
         template <int dim>
-
-
-
-
         StokesPreconditioner<dim>::
         StokesPreconditioner (const FiniteElement<dim> &finite_element,
                               const Quadrature<dim>    &quadrature,
@@ -466,393 +438,6 @@ namespace aspect
   }
 
 
-  /**
-   * Compute the variation in the entropy needed in the definition of the
-   * artificial viscosity used to stabilize the composition/temperature equation.
-   */
-  template <int dim>
-  double
-  Simulator<dim>::get_entropy_variation (const double average_field,
-                                         const AdvectionField &advection_field) const
-  {
-    // only do this if we really need entropy
-    // variation. otherwise return something that's obviously
-    // nonsensical
-    if (parameters.stabilization_alpha != 2)
-      return std::numeric_limits<double>::quiet_NaN();
-
-    // record maximal entropy on Gauss quadrature
-    // points
-    const QGauss<dim> quadrature_formula (parameters.temperature_degree+1);
-    const unsigned int n_q_points = quadrature_formula.size();
-
-    const FEValuesExtractors::Scalar field = advection_field.scalar_extractor(introspection);
-
-    FEValues<dim> fe_values (finite_element, quadrature_formula,
-                             update_values | update_JxW_values);
-    std::vector<double> old_field_values(n_q_points);
-    std::vector<double> old_old_field_values(n_q_points);
-
-    double min_entropy = std::numeric_limits<double>::max(),
-           max_entropy = -std::numeric_limits<double>::max(),
-           area = 0,
-           entropy_integrated = 0;
-
-    // loop over all locally owned cells and evaluate the entropy
-    // at all quadrature points. keep a running tally of the
-    // integral over the entropy as well as the area and the
-    // maximal and minimal entropy
-    typename DoFHandler<dim>::active_cell_iterator
-    cell = dof_handler.begin_active(),
-    endc = dof_handler.end();
-    for (; cell!=endc; ++cell)
-      if (cell->is_locally_owned())
-        {
-          fe_values.reinit (cell);
-          fe_values[field].get_function_values (old_solution,
-                                                old_field_values);
-          fe_values[field].get_function_values (old_old_solution,
-                                                old_old_field_values);
-          for (unsigned int q=0; q<n_q_points; ++q)
-            {
-              const double T = (old_field_values[q] +
-                                old_old_field_values[q]) / 2;
-              const double entropy = ((T-average_field) *
-                                      (T-average_field));
-
-              min_entropy = std::min (min_entropy, entropy);
-              max_entropy = std::max (max_entropy, entropy);
-
-              area += fe_values.JxW(q);
-              entropy_integrated += fe_values.JxW(q) * entropy;
-            }
-        }
-
-    // do MPI data exchange: we need to sum over
-    // the two integrals (area,
-    // entropy_integrated), and get the extrema
-    // for maximum and minimum. combine
-    // MPI_Allreduce for two values since that is
-    // an expensive operation
-    const double local_for_sum[2] = { entropy_integrated, area },
-                                    local_for_max[2] = { -min_entropy, max_entropy };
-    double global_for_sum[2], global_for_max[2];
-
-    dealii::Utilities::MPI::sum (local_for_sum, mpi_communicator, global_for_sum);
-    dealii::Utilities::MPI::max (local_for_max, mpi_communicator, global_for_max);
-
-    const double average_entropy = global_for_sum[0] / global_for_sum[1];
-
-    // return the maximal deviation of the entropy everywhere from the
-    // average value
-    return std::max(global_for_max[1] - average_entropy,
-                    average_entropy - (-global_for_max[0]));
-  }
-
-
-  template <int dim>
-  double
-  Simulator<dim>::
-  compute_viscosity (internal::Assembly::Scratch::AdvectionSystem<dim> &scratch,
-                     const double                        global_u_infty,
-                     const double                        global_field_variation,
-                     const double                        average_field,
-                     const double                        global_entropy_variation,
-                     const double                        cell_diameter,
-                     const AdvectionField               &advection_field) const
-  {
-    // discontinuous Galerkin doesn't require an artificial viscosity
-    if (advection_field.is_discontinuous(introspection))
-      return 0.;
-
-    std::vector<double> residual = assemblers->compute_advection_system_residual(*scratch.material_model_inputs.cell,
-                                                                                 advection_field,
-                                                                                 scratch);
-
-    double max_residual = 0;
-    double max_velocity = 0;
-    double max_density = (advection_field.is_temperature()) ? 0.0 : 1.0;
-    double max_specific_heat = (advection_field.is_temperature()) ? 0.0 : 1.0;
-    double max_conductivity = 0;
-
-    for (unsigned int q=0; q < scratch.finite_element_values.n_quadrature_points; ++q)
-      {
-        const Tensor<1,dim> velocity = (scratch.old_velocity_values[q] +
-                                        scratch.old_old_velocity_values[q]) / 2;
-
-        const double strain_rate = ((scratch.old_strain_rates[q]
-                                     + scratch.old_old_strain_rates[q]) / 2).norm();
-
-        if (parameters.stabilization_alpha == 2)
-          {
-            const double field = (scratch.old_field_values[q] + scratch.old_old_field_values[q]) / 2;
-            residual[q] *= std::abs(field - average_field);
-          }
-
-        max_residual = std::max (residual[q],     max_residual);
-        max_velocity = std::max (velocity.norm()
-                                 + parameters.stabilization_gamma * strain_rate * cell_diameter,
-                                 max_velocity);
-
-        if (advection_field.is_temperature())
-          {
-            max_density = std::max       (scratch.material_model_outputs.densities[q],              max_density);
-            max_specific_heat = std::max (scratch.material_model_outputs.specific_heat[q],          max_specific_heat);
-            max_conductivity = std::max  (scratch.material_model_outputs.thermal_conductivities[q], max_conductivity);
-          }
-      }
-
-    // If the velocity is 0 we have to assume a sensible velocity to calculate
-    // an artificial diffusion. We choose similar to nondimensional
-    // formulations: v ~ thermal_diffusivity / length_scale, which cancels
-    // the density and specific heat from the entropy formulation. It seems
-    // surprising at first that only the conductivity remains, but remember
-    // that this actually *is* an additional artificial diffusion.
-    if (std::abs(global_u_infty) < 1e-50)
-      return parameters.stabilization_beta *
-             max_conductivity / geometry_model->length_scale() *
-             cell_diameter;
-
-    const double max_viscosity = parameters.stabilization_beta *
-                                 max_density *
-                                 max_specific_heat *
-                                 max_velocity * cell_diameter;
-
-    if (timestep_number <= 1
-        || std::abs(global_entropy_variation) < 1e-50
-        || std::abs(global_field_variation) < 1e-50)
-      // we don't have sensible timesteps during the first two iterations
-      // and we can not divide by the entropy_variation if it is zero
-      return max_viscosity;
-    else
-      {
-        Assert (old_time_step > 0, ExcInternalError());
-
-        double entropy_viscosity;
-        if (parameters.stabilization_alpha == 2)
-          entropy_viscosity = (parameters.stabilization_c_R *
-                               cell_diameter * cell_diameter *
-                               max_residual /
-                               global_entropy_variation);
-        else
-          entropy_viscosity = (parameters.stabilization_c_R *
-                               cell_diameter * global_Omega_diameter *
-                               max_velocity * max_residual /
-                               (global_u_infty * global_field_variation));
-
-
-        return std::min (max_viscosity, entropy_viscosity);
-      }
-  }
-
-  template <int dim>
-  template <typename T>
-  void
-  Simulator<dim>::
-  get_artificial_viscosity (Vector<T> &viscosity_per_cell,
-                            const AdvectionField &advection_field) const
-  {
-    Assert(viscosity_per_cell.size()==triangulation.n_active_cells(), ExcInternalError());
-
-    if (advection_field.field_type == AdvectionField::compositional_field)
-      Assert(parameters.n_compositional_fields > advection_field.compositional_variable, ExcInternalError());
-
-    viscosity_per_cell = 0.0;
-
-    //discontinuous Galerkin doesn't require an artificial viscosity
-    if (advection_field.is_discontinuous(introspection))
-      return;
-
-    const std::pair<double,double>
-    global_field_range = get_extrapolated_advection_field_range (advection_field);
-    double global_entropy_variation = get_entropy_variation ((global_field_range.first +
-                                                              global_field_range.second) / 2,
-                                                             advection_field);
-    double global_max_velocity = get_maximal_velocity(old_solution);
-
-    const UpdateFlags update_flags = update_values |
-                                     update_gradients |
-                                     (advection_field.is_temperature() ? update_hessians : update_default) |
-                                     update_quadrature_points |
-                                     update_JxW_values;
-
-    /* Because we can only get here in the continuous case, which never requires
-     * face integrals, we supply no face_update_flags and an invalid
-     * face_quadrature to the scratch object to reduce the initialization cost.
-     */
-    const UpdateFlags face_update_flags = update_default;
-
-    internal::Assembly::Scratch::
-    AdvectionSystem<dim> scratch (finite_element,
-                                  finite_element.base_element(advection_field.base_element(introspection)),
-                                  *mapping,
-                                  QGauss<dim>(advection_field.polynomial_degree(introspection)
-                                              +
-                                              (parameters.stokes_velocity_degree+1)/2),
-                                  Quadrature<dim-1> (),
-                                  update_flags,
-                                  face_update_flags,
-                                  parameters.n_compositional_fields);
-
-    typename DoFHandler<dim>::active_cell_iterator cell = dof_handler.begin_active();
-    for (; cell<dof_handler.end(); ++cell)
-      {
-        if (!cell->is_locally_owned()
-            || (parameters.use_artificial_viscosity_smoothing  == true  &&  cell->is_artificial()))
-          {
-            viscosity_per_cell[cell->active_cell_index()]=-1;
-            continue;
-          }
-
-        const unsigned int n_q_points    = scratch.finite_element_values.n_quadrature_points;
-
-        // also have the number of dofs that correspond just to the element for
-        // the system we are currently trying to assemble
-        const unsigned int advection_dofs_per_cell = scratch.phi_field.size();
-        (void)advection_dofs_per_cell;
-        Assert (advection_dofs_per_cell < scratch.finite_element_values.get_fe().dofs_per_cell, ExcInternalError());
-        Assert (scratch.grad_phi_field.size() == advection_dofs_per_cell, ExcInternalError());
-        Assert (scratch.phi_field.size() == advection_dofs_per_cell, ExcInternalError());
-
-        const FEValuesExtractors::Scalar solution_field = advection_field.scalar_extractor(introspection);
-
-        scratch.finite_element_values.reinit (cell);
-
-        // get all dof indices on the current cell, then extract those
-        // that correspond to the solution_field we are interested in
-        cell->get_dof_indices (scratch.local_dof_indices);
-
-        // initialize all of the scratch fields for further down
-        scratch.finite_element_values[introspection.extractors.temperature].get_function_values (old_solution,
-            scratch.old_temperature_values);
-        scratch.finite_element_values[introspection.extractors.temperature].get_function_values (old_old_solution,
-            scratch.old_old_temperature_values);
-
-        scratch.finite_element_values[introspection.extractors.velocities].get_function_symmetric_gradients (old_solution,
-            scratch.old_strain_rates);
-        scratch.finite_element_values[introspection.extractors.velocities].get_function_symmetric_gradients (old_old_solution,
-            scratch.old_old_strain_rates);
-
-        scratch.finite_element_values[introspection.extractors.pressure].get_function_values (old_solution,
-            scratch.old_pressure);
-        scratch.finite_element_values[introspection.extractors.pressure].get_function_values (old_old_solution,
-            scratch.old_old_pressure);
-
-        for (unsigned int c=0; c<parameters.n_compositional_fields; ++c)
-          {
-            scratch.finite_element_values[introspection.extractors.compositional_fields[c]].get_function_values(old_solution,
-                scratch.old_composition_values[c]);
-            scratch.finite_element_values[introspection.extractors.compositional_fields[c]].get_function_values(old_old_solution,
-                scratch.old_old_composition_values[c]);
-          }
-
-        scratch.finite_element_values[introspection.extractors.velocities].get_function_values (old_solution,
-            scratch.old_velocity_values);
-        scratch.finite_element_values[introspection.extractors.velocities].get_function_values (old_old_solution,
-            scratch.old_old_velocity_values);
-        scratch.finite_element_values[introspection.extractors.velocities].get_function_values(current_linearization_point,
-            scratch.current_velocity_values);
-
-        scratch.finite_element_values[introspection.extractors.pressure].get_function_gradients (old_solution,
-            scratch.old_pressure_gradients);
-        scratch.finite_element_values[introspection.extractors.pressure].get_function_gradients (old_old_solution,
-            scratch.old_old_pressure_gradients);
-
-
-        scratch.old_field_values = (advection_field.is_temperature()
-                                    ?
-                                    scratch.old_temperature_values
-                                    :
-                                    scratch.old_composition_values[advection_field.compositional_variable]);
-        scratch.old_old_field_values = (advection_field.is_temperature()
-                                        ?
-                                        scratch.old_old_temperature_values
-                                        :
-                                        scratch.old_old_composition_values[advection_field.compositional_variable]);
-
-        scratch.finite_element_values[solution_field].get_function_gradients (old_solution,
-                                                                              scratch.old_field_grads);
-        scratch.finite_element_values[solution_field].get_function_gradients (old_old_solution,
-                                                                              scratch.old_old_field_grads);
-
-        if (advection_field.is_temperature())
-          {
-            scratch.finite_element_values[solution_field].get_function_laplacians (old_solution,
-                                                                                   scratch.old_field_laplacians);
-            scratch.finite_element_values[solution_field].get_function_laplacians (old_old_solution,
-                                                                                   scratch.old_old_field_laplacians);
-          }
-
-        if (parameters.include_melt_transport && melt_handler->is_porosity(advection_field))
-          {
-            scratch.finite_element_values[introspection.extractors.velocities].get_function_divergences (current_linearization_point,
-                scratch.current_velocity_divergences);
-          }
-
-        /**
-         * Explicit material model inputs and outputs.
-         */
-        for (unsigned int q=0; q<n_q_points; ++q)
-          {
-            scratch.material_model_inputs.temperature[q] = (scratch.old_temperature_values[q] + scratch.old_old_temperature_values[q]) / 2;
-            scratch.material_model_inputs.position[q] = scratch.finite_element_values.quadrature_point(q);
-            scratch.material_model_inputs.pressure[q] = (scratch.old_pressure[q] + scratch.old_old_pressure[q]) / 2;
-            scratch.material_model_inputs.velocity[q] = (scratch.old_velocity_values[q] + scratch.old_old_velocity_values[q]) / 2;
-            scratch.material_model_inputs.pressure_gradient[q] = (scratch.old_pressure_gradients[q] + scratch.old_old_pressure_gradients[q]) / 2;
-
-            for (unsigned int c=0; c<parameters.n_compositional_fields; ++c)
-              scratch.material_model_inputs.composition[q][c] = (scratch.old_composition_values[c][q] + scratch.old_old_composition_values[c][q]) / 2;
-            scratch.material_model_inputs.strain_rate[q] = (scratch.old_strain_rates[q] + scratch.old_old_strain_rates[q]) / 2;
-          }
-        scratch.material_model_inputs.cell = &cell;
-        create_additional_material_model_outputs(scratch.material_model_outputs);
-
-        material_model->evaluate(scratch.material_model_inputs,scratch.material_model_outputs);
-        MaterialModel::MaterialAveraging::average (parameters.material_averaging,
-                                                   cell,
-                                                   scratch.finite_element_values.get_quadrature(),
-                                                   scratch.finite_element_values.get_mapping(),
-                                                   scratch.material_model_outputs);
-
-        viscosity_per_cell[cell->active_cell_index()] = compute_viscosity(scratch,
-                                                                          global_max_velocity,
-                                                                          global_field_range.second - global_field_range.first,
-                                                                          0.5 * (global_field_range.second + global_field_range.first),
-                                                                          global_entropy_variation,
-                                                                          cell->diameter(),
-                                                                          advection_field);
-      }
-
-    // if set to true, the maximum of the artificial viscosity in the cell as well
-    // as the neighbors of the cell is computed and used instead
-    if (parameters.use_artificial_viscosity_smoothing  == true)
-      {
-        Vector<T> viscosity_per_cell_temp;
-        viscosity_per_cell_temp.reinit(triangulation.n_active_cells());
-
-        viscosity_per_cell_temp = viscosity_per_cell;
-        typename DoFHandler<dim>::active_cell_iterator
-        cell,
-        end_cell = dof_handler.end();
-        for (cell = dof_handler.begin_active(); cell!=end_cell; ++cell)
-          {
-            if (cell->is_locally_owned())
-              for (unsigned int face_no=0; face_no<GeometryInfo<dim>::faces_per_cell; ++face_no)
-                if (cell->at_boundary(face_no) == false)
-                  {
-                    if (cell->neighbor(face_no)->active())
-                      viscosity_per_cell[cell->active_cell_index()] = std::max(viscosity_per_cell[cell->active_cell_index()],
-                                                                               viscosity_per_cell_temp[cell->neighbor(face_no)->active_cell_index()]);
-                    else
-                      for (unsigned int l=0; l<cell->neighbor(face_no)->n_children(); l++)
-                        if (cell->neighbor(face_no)->child(l)->active())
-                          viscosity_per_cell[cell->active_cell_index()] = std::max(viscosity_per_cell[cell->active_cell_index()],
-                                                                                   viscosity_per_cell_temp[cell->neighbor(face_no)->child(l)->active_cell_index()]);
-                  }
-          }
-      }
-  }
-
 
 
   template <int dim>
@@ -909,1310 +494,6 @@ namespace aspect
 
   namespace Assemblers
   {
-
-    /**
-     * A class for the definition of functions that implement the
-     * linear system terms for the *complete* compressible or
-     * incompressible equations.
-     */
-    template <int dim>
-    class CompleteEquations : public aspect::internal::Assembly::Assemblers::AssemblerBase<dim>,
-      public SimulatorAccess<dim>
-    {
-      public:
-        void
-        local_assemble_stokes_preconditioner (const double                                             pressure_scaling,
-                                              internal::Assembly::Scratch::StokesPreconditioner<dim>  &scratch,
-                                              internal::Assembly::CopyData::StokesPreconditioner<dim> &data) const
-        {
-          const Introspection<dim> &introspection = this->introspection();
-          const FiniteElement<dim> &fe = this->get_fe();
-          const unsigned int stokes_dofs_per_cell = data.local_dof_indices.size();
-          const unsigned int n_q_points           = scratch.finite_element_values.n_quadrature_points;
-
-          // First loop over all dofs and find those that are in the Stokes system
-          // save the component (pressure and dim velocities) each belongs to.
-          for (unsigned int i = 0, i_stokes = 0; i_stokes < stokes_dofs_per_cell; /*increment at end of loop*/)
-            {
-              if (introspection.is_stokes_component(fe.system_to_component_index(i).first))
-                {
-                  scratch.dof_component_indices[i_stokes] = fe.system_to_component_index(i).first;
-                  ++i_stokes;
-                }
-              ++i;
-            }
-
-          // Loop over all quadrature points and assemble their contributions to
-          // the preconditioner matrix
-          for (unsigned int q = 0; q < n_q_points; ++q)
-            {
-              for (unsigned int i = 0, i_stokes = 0; i_stokes < stokes_dofs_per_cell; /*increment at end of loop*/)
-                {
-                  if (introspection.is_stokes_component(fe.system_to_component_index(i).first))
-                    {
-                      scratch.grads_phi_u[i_stokes] =
-                        scratch.finite_element_values[introspection.extractors
-                                                      .velocities].symmetric_gradient(i, q);
-                      scratch.phi_p[i_stokes] = scratch.finite_element_values[introspection
-                                                                              .extractors.pressure].value(i, q);
-                      ++i_stokes;
-                    }
-                  ++i;
-                }
-
-              const double eta = scratch.material_model_outputs.viscosities[q];
-              const double one_over_eta = 1. / eta;
-
-              const SymmetricTensor<4, dim> &stress_strain_director = scratch
-                                                                      .material_model_outputs.stress_strain_directors[q];
-              const bool use_tensor = (stress_strain_director
-                                       != dealii::identity_tensor<dim>());
-
-              const double JxW = scratch.finite_element_values.JxW(q);
-
-              for (unsigned int i = 0; i < stokes_dofs_per_cell; ++i)
-                for (unsigned int j = 0; j < stokes_dofs_per_cell; ++j)
-                  if (scratch.dof_component_indices[i] ==
-                      scratch.dof_component_indices[j])
-                    data.local_matrix(i, j) += ((
-                                                  use_tensor ?
-                                                  eta * (scratch.grads_phi_u[i]
-                                                         * stress_strain_director
-                                                         * scratch.grads_phi_u[j]) :
-                                                  eta * (scratch.grads_phi_u[i]
-                                                         * scratch.grads_phi_u[j]))
-                                                + one_over_eta * pressure_scaling
-                                                * pressure_scaling
-                                                * (scratch.phi_p[i] * scratch
-                                                   .phi_p[j]))
-                                               * JxW;
-            }
-        }
-
-
-        void
-        local_assemble_stokes_system_compressible (const double                                     pressure_scaling,
-                                                   const bool                                       rebuild_stokes_matrix,
-                                                   internal::Assembly::Scratch::StokesSystem<dim>  &scratch,
-                                                   internal::Assembly::CopyData::StokesSystem<dim> &data) const
-        {
-          const Introspection<dim> &introspection = this->introspection();
-          const FiniteElement<dim> &fe = this->get_fe();
-          const unsigned int stokes_dofs_per_cell = data.local_dof_indices.size();
-          const unsigned int n_q_points    = scratch.finite_element_values.n_quadrature_points;
-
-          for (unsigned int q=0; q<n_q_points; ++q)
-            {
-              for (unsigned int i=0, i_stokes=0; i_stokes<stokes_dofs_per_cell; /*increment at end of loop*/)
-                {
-                  if (introspection.is_stokes_component(fe.system_to_component_index(i).first))
-                    {
-                      scratch.phi_u[i_stokes] = scratch.finite_element_values[introspection.extractors.velocities].value (i,q);
-                      scratch.phi_p[i_stokes] = scratch.finite_element_values[introspection.extractors.pressure].value (i, q);
-                      if (rebuild_stokes_matrix)
-                        {
-                          scratch.grads_phi_u[i_stokes] = scratch.finite_element_values[introspection.extractors.velocities].symmetric_gradient(i,q);
-                          scratch.div_phi_u[i_stokes]   = scratch.finite_element_values[introspection.extractors.velocities].divergence (i, q);
-                        }
-                      ++i_stokes;
-                    }
-                  ++i;
-                }
-
-              // Viscosity scalar
-              const double eta = (rebuild_stokes_matrix
-                                  ?
-                                  scratch.material_model_outputs.viscosities[q]
-                                  :
-                                  std::numeric_limits<double>::quiet_NaN());
-              const double eta_two_thirds = (rebuild_stokes_matrix
-                                             ?
-                                             scratch.material_model_outputs.viscosities[q] * 2.0 / 3.0
-                                             :
-                                             std::numeric_limits<double>::quiet_NaN());
-
-              const SymmetricTensor<4,dim> &stress_strain_director =
-                scratch.material_model_outputs.stress_strain_directors[q];
-              const bool use_tensor = (stress_strain_director !=  dealii::identity_tensor<dim> ());
-
-              const Tensor<1,dim>
-              gravity = this->get_gravity_model().gravity_vector (scratch.finite_element_values.quadrature_point(q));
-
-              const double compressibility
-                = scratch.material_model_outputs.compressibilities[q];
-              const double density = scratch.material_model_outputs.densities[q];
-
-              const double JxW = scratch.finite_element_values.JxW(q);
-
-              for (unsigned int i=0; i<stokes_dofs_per_cell; ++i)
-                {
-                  data.local_rhs(i) += (
-                                         (density * gravity * scratch.phi_u[i])
-                                         +
-                                         // add the term that results from the compressibility. compared
-                                         // to the manual, this term seems to have the wrong sign, but this
-                                         // is because we negate the entire equation to make sure we get
-                                         // -div(u) as the adjoint operator of grad(p) (see below where
-                                         // we assemble the matrix)
-                                         (pressure_scaling *
-                                          compressibility * density *
-                                          (scratch.velocity_values[q] * gravity) *
-                                          scratch.phi_p[i])
-                                       )
-                                       * JxW;
-
-                  if (rebuild_stokes_matrix)
-                    for (unsigned int j=0; j<stokes_dofs_per_cell; ++j)
-                      {
-                        data.local_matrix(i,j) += ( (use_tensor ?
-                                                     eta * 2.0 * (scratch.grads_phi_u[i] * stress_strain_director * scratch.grads_phi_u[j])
-                                                     :
-                                                     eta * 2.0 * (scratch.grads_phi_u[i] * scratch.grads_phi_u[j]))
-                                                    - (use_tensor ?
-                                                       eta_two_thirds * (scratch.div_phi_u[i] * trace(stress_strain_director * scratch.grads_phi_u[j]))
-                                                       :
-                                                       eta_two_thirds * (scratch.div_phi_u[i] * scratch.div_phi_u[j])
-                                                      )
-                                                    - (pressure_scaling *
-                                                       scratch.div_phi_u[i] * scratch.phi_p[j])
-                                                    // finally the term -div(u). note the negative sign to make this
-                                                    // operator adjoint to the grad(p) term
-                                                    - (pressure_scaling *
-                                                       scratch.phi_p[i] * scratch.div_phi_u[j]))
-                                                  * JxW;
-                      }
-                }
-            }
-
-        }
-
-
-        void
-        local_assemble_stokes_system_incompressible (const double                                     pressure_scaling,
-                                                     const bool                                       rebuild_stokes_matrix,
-                                                     internal::Assembly::Scratch::StokesSystem<dim>  &scratch,
-                                                     internal::Assembly::CopyData::StokesSystem<dim> &data) const
-        {
-          const Introspection<dim> &introspection = this->introspection();
-          const FiniteElement<dim> &fe = this->get_fe();
-          const unsigned int stokes_dofs_per_cell = data.local_dof_indices.size();
-          const unsigned int n_q_points    = scratch.finite_element_values.n_quadrature_points;
-
-          for (unsigned int q=0; q<n_q_points; ++q)
-            {
-              for (unsigned int i=0, i_stokes=0; i_stokes<stokes_dofs_per_cell; /*increment at end of loop*/)
-                {
-                  if (introspection.is_stokes_component(fe.system_to_component_index(i).first))
-                    {
-                      scratch.phi_u[i_stokes] = scratch.finite_element_values[introspection.extractors.velocities].value (i,q);
-                      scratch.phi_p[i_stokes] = scratch.finite_element_values[introspection.extractors.pressure].value (i, q);
-                      if (rebuild_stokes_matrix)
-                        {
-                          scratch.grads_phi_u[i_stokes] = scratch.finite_element_values[introspection.extractors.velocities].symmetric_gradient(i,q);
-                          scratch.div_phi_u[i_stokes]   = scratch.finite_element_values[introspection.extractors.velocities].divergence (i, q);
-                        }
-                      ++i_stokes;
-                    }
-                  ++i;
-                }
-
-
-              // Viscosity scalar
-              const double eta = (rebuild_stokes_matrix
-                                  ?
-                                  scratch.material_model_outputs.viscosities[q]
-                                  :
-                                  std::numeric_limits<double>::quiet_NaN());
-
-              const SymmetricTensor<4,dim> &stress_strain_director =
-                scratch.material_model_outputs.stress_strain_directors[q];
-              const bool use_tensor = (stress_strain_director !=  dealii::identity_tensor<dim> ());
-
-              const Tensor<1,dim>
-              gravity = this->get_gravity_model().gravity_vector (scratch.finite_element_values.quadrature_point(q));
-
-              const double density = scratch.material_model_outputs.densities[q];
-
-              const double JxW = scratch.finite_element_values.JxW(q);
-
-              for (unsigned int i=0; i<stokes_dofs_per_cell; ++i)
-                {
-                  data.local_rhs(i) += (density * gravity * scratch.phi_u[i])
-                                       * JxW;
-
-                  if (rebuild_stokes_matrix)
-                    for (unsigned int j=0; j<stokes_dofs_per_cell; ++j)
-                      {
-                        data.local_matrix(i,j) += ( (use_tensor ?
-                                                     eta * 2.0 * (scratch.grads_phi_u[i] * stress_strain_director * scratch.grads_phi_u[j])
-                                                     :
-                                                     eta * 2.0 * (scratch.grads_phi_u[i] * scratch.grads_phi_u[j]))
-                                                    - (pressure_scaling *
-                                                       scratch.div_phi_u[i] * scratch.phi_p[j])
-                                                    // finally the term -div(u). note the negative sign to make this
-                                                    // operator adjoint to the grad(p) term
-                                                    - (pressure_scaling *
-                                                       scratch.phi_p[i] * scratch.div_phi_u[j]))
-                                                  * JxW;
-                      }
-                }
-            }
-        }
-
-        void
-        local_assemble_advection_system (const typename Simulator<dim>::AdvectionField &advection_field,
-                                         const double artificial_viscosity,
-                                         internal::Assembly::Scratch::AdvectionSystem<dim>  &scratch,
-                                         internal::Assembly::CopyData::AdvectionSystem<dim> &data) const
-        {
-          const Introspection<dim> &introspection = this->introspection();
-          const FiniteElement<dim> &fe = this->get_fe();
-          const unsigned int n_q_points = scratch.finite_element_values.n_quadrature_points;
-          const unsigned int advection_dofs_per_cell = data.local_dof_indices.size();
-
-          const bool   use_bdf2_scheme = (this->get_timestep_number() > 1);
-          const double time_step = this->get_timestep();
-          const double old_time_step = this->get_old_timestep();
-
-          const double bdf2_factor = (use_bdf2_scheme)? ((2*time_step + old_time_step) /
-                                                         (time_step + old_time_step)) : 1.0;
-
-          const bool advection_field_is_temperature = advection_field.is_temperature();
-          const unsigned int solution_component = advection_field.component_index(introspection);
-
-          const FEValuesExtractors::Scalar solution_field = advection_field.scalar_extractor(introspection);
-
-          for (unsigned int q=0; q<n_q_points; ++q)
-            {
-              // precompute the values of shape functions and their gradients.
-              // We only need to look up values of shape functions if they
-              // belong to 'our' component. They are zero otherwise anyway.
-              // Note that we later only look at the values that we do set here.
-              for (unsigned int i=0, i_advection=0; i_advection<advection_dofs_per_cell;/*increment at end of loop*/)
-                {
-                  if (fe.system_to_component_index(i).first == solution_component)
-                    {
-                      scratch.grad_phi_field[i_advection] = scratch.finite_element_values[solution_field].gradient (i,q);
-                      scratch.phi_field[i_advection]      = scratch.finite_element_values[solution_field].value (i,q);
-                      ++i_advection;
-                    }
-                  ++i;
-                }
-
-              const double density_c_P              =
-                ((advection_field_is_temperature)
-                 ?
-                 scratch.material_model_outputs.densities[q] *
-                 scratch.material_model_outputs.specific_heat[q]
-                 :
-                 1.0);
-
-              Assert (density_c_P >= 0,
-                      ExcMessage ("The product of density and c_P needs to be a "
-                                  "non-negative quantity."));
-
-              const double conductivity =
-                ((advection_field_is_temperature)
-                 ?
-                 scratch.material_model_outputs.thermal_conductivities[q]
-                 :
-                 0.0);
-              const double latent_heat_LHS =
-                ((advection_field_is_temperature)
-                 ?
-                 scratch.heating_model_outputs.lhs_latent_heat_terms[q]
-                 :
-                 0.0);
-              Assert (density_c_P + latent_heat_LHS >= 0,
-                      ExcMessage ("The sum of density times c_P and the latent heat contribution "
-                                  "to the left hand side needs to be a non-negative quantity."));
-
-              const double gamma =
-                ((advection_field_is_temperature)
-                 ?
-                 scratch.heating_model_outputs.heating_source_terms[q]
-                 :
-                 0.0);
-
-              const double reaction_term =
-                ((advection_field_is_temperature)
-                 ?
-                 0.0
-                 :
-                 scratch.material_model_outputs.reaction_terms[q][advection_field.compositional_variable]);
-
-              const double field_term_for_rhs
-                = (use_bdf2_scheme ?
-                   (scratch.old_field_values[q] *
-                    (1 + time_step/old_time_step)
-                    -
-                    scratch.old_old_field_values[q] *
-                    (time_step * time_step) /
-                    (old_time_step * (time_step + old_time_step)))
-                   :
-                   scratch.old_field_values[q])
-                  *
-                  (density_c_P + latent_heat_LHS);
-
-              Tensor<1,dim> current_u = scratch.current_velocity_values[q];
-              //Subtract off the mesh velocity for ALE corrections if necessary
-              if (this->get_parameters().free_surface_enabled)
-                current_u -= scratch.mesh_velocity_values[q];
-
-              const double JxW = scratch.finite_element_values.JxW(q);
-
-              // do the actual assembly. note that we only need to loop over the advection
-              // shape functions because these are the only contributions we compute here
-              for (unsigned int i=0; i<advection_dofs_per_cell; ++i)
-                {
-                  data.local_rhs(i)
-                  += (field_term_for_rhs * scratch.phi_field[i]
-                      + time_step *
-                      scratch.phi_field[i]
-                      * gamma
-                      + scratch.phi_field[i]
-                      * reaction_term)
-                     *
-                     JxW;
-
-                  for (unsigned int j=0; j<advection_dofs_per_cell; ++j)
-                    {
-                      data.local_matrix(i,j)
-                      += (
-                           (time_step * (conductivity + artificial_viscosity)
-                            * (scratch.grad_phi_field[i] * scratch.grad_phi_field[j]))
-                           + ((time_step * (scratch.phi_field[i] * (current_u * scratch.grad_phi_field[j])))
-                              + (bdf2_factor * scratch.phi_field[i] * scratch.phi_field[j])) *
-                           (density_c_P + latent_heat_LHS)
-                         )
-                         * JxW;
-                    }
-                }
-            }
-        }
-
-
-        std::vector<double>
-        compute_advection_system_residual(const typename Simulator<dim>::AdvectionField     &advection_field,
-                                          internal::Assembly::Scratch::AdvectionSystem<dim> &scratch) const
-        {
-          const unsigned int n_q_points = scratch.finite_element_values.n_quadrature_points;
-          std::vector<double> residuals(n_q_points);
-
-          if (advection_field.is_temperature())
-            this->get_heating_model_manager().evaluate(scratch.material_model_inputs,
-                                                       scratch.material_model_outputs,
-                                                       scratch.heating_model_outputs);
-
-          for (unsigned int q=0; q < n_q_points; ++q)
-            {
-              const Tensor<1,dim> u = (scratch.old_velocity_values[q] +
-                                       scratch.old_old_velocity_values[q]) / 2;
-
-              const double dField_dt = (this->get_old_timestep() == 0.0) ? 0.0 :
-                                       (
-                                         ((scratch.old_field_values)[q] - (scratch.old_old_field_values)[q])
-                                         / this->get_old_timestep());
-              const double u_grad_field = u * (scratch.old_field_grads[q] +
-                                               scratch.old_old_field_grads[q]) / 2;
-
-              if (advection_field.is_temperature())
-                {
-                  const double density       = scratch.material_model_outputs.densities[q];
-                  const double conductivity  = scratch.material_model_outputs.thermal_conductivities[q];
-                  const double c_P           = scratch.material_model_outputs.specific_heat[q];
-                  const double k_Delta_field = conductivity * (scratch.old_field_laplacians[q] +
-                                                               scratch.old_old_field_laplacians[q]) / 2;
-
-                  const double gamma           = scratch.heating_model_outputs.heating_source_terms[q];
-                  const double latent_heat_LHS = scratch.heating_model_outputs.lhs_latent_heat_terms[q];
-
-                  residuals[q]
-                    = std::abs((density * c_P + latent_heat_LHS) * (dField_dt + u_grad_field) - k_Delta_field - gamma);
-                }
-              else
-                {
-                  const double dreaction_term_dt = (this->get_old_timestep() == 0) ? 0.0 :
-                                                   scratch.material_model_outputs.reaction_terms[q][advection_field.compositional_variable]
-                                                   / this->get_old_timestep();
-
-                  residuals[q] = std::abs(dField_dt + u_grad_field - dreaction_term_dt);
-                }
-            }
-          return residuals;
-        }
-
-
-        void
-        local_assemble_discontinuous_advection_boundary_face_terms(const typename DoFHandler<dim>::active_cell_iterator &cell,
-                                                                   const unsigned int face_no,
-                                                                   const typename Simulator<dim>::AdvectionField &advection_field,
-                                                                   internal::Assembly::Scratch::AdvectionSystem<dim> &scratch,
-                                                                   internal::Assembly::CopyData::AdvectionSystem<dim> &data) const
-        {
-          const Parameters<dim> &parameters = this->get_parameters();
-          const Introspection<dim> &introspection = this->introspection();
-          const FiniteElement<dim> &fe = this->get_fe();
-
-          const unsigned int n_q_points    = scratch.face_finite_element_values->n_quadrature_points;
-
-          const double time_step = this->get_timestep();
-
-          if (!advection_field.is_discontinuous(introspection))
-            return;
-
-          // also have the number of dofs that correspond just to the element for
-          // the system we are currently trying to assemble
-          const unsigned int advection_dofs_per_cell = data.local_dof_indices.size();
-
-          Assert (advection_dofs_per_cell < scratch.face_finite_element_values->get_fe().dofs_per_cell, ExcInternalError());
-          Assert (scratch.face_grad_phi_field.size() == advection_dofs_per_cell, ExcInternalError());
-          Assert (scratch.face_phi_field.size() == advection_dofs_per_cell, ExcInternalError());
-
-          const unsigned int solution_component = advection_field.component_index(introspection);
-
-          const FEValuesExtractors::Scalar solution_field = advection_field.scalar_extractor(introspection);
-
-          typename DoFHandler<dim>::face_iterator face = cell->face (face_no);
-
-          if (((parameters.fixed_temperature_boundary_indicators.find(
-                  face->boundary_id()
-                )
-                != parameters.fixed_temperature_boundary_indicators.end())
-               && (advection_field.is_temperature()))
-              ||
-              (( parameters.fixed_composition_boundary_indicators.find(
-                   face->boundary_id()
-                 )
-                 != parameters.fixed_composition_boundary_indicators.end())
-               && (!advection_field.is_temperature())))
-            {
-              /*
-               * We are in the case of a Dirichlet temperature or composition boundary.
-               * In the temperature case, impose the Dirichlet value weakly using a matrix term
-               * and RHS term. In the composition case, Dirichlet conditions can only be imposed
-               * on inflow boundaries, and we only have the flow-dependent terms, so we only
-               * assemble the corresponding flow-dependent and matrix and RHS terms
-               * if we are on an inflow boundary.
-               */
-
-              for (unsigned int q=0; q<n_q_points; ++q)
-                {
-                  // precompute the values of shape functions and their gradients.
-                  // We only need to look up values of shape functions if they
-                  // belong to 'our' component. They are zero otherwise anyway.
-                  // Note that we later only look at the values that we do set here.
-                  for (unsigned int i=0, i_advection=0; i_advection<advection_dofs_per_cell; /*increment at end of loop*/)
-                    {
-                      if (fe.system_to_component_index(i).first == solution_component)
-                        {
-                          scratch.face_grad_phi_field[i_advection] = (*scratch.face_finite_element_values)[solution_field].gradient (i, q);
-                          scratch.face_phi_field[i_advection]      = (*scratch.face_finite_element_values)[solution_field].value (i, q);
-                          ++i_advection;
-                        }
-                      ++i;
-                    }
-
-                  const double density_c_P              =
-                    ((advection_field.is_temperature())
-                     ?
-                     scratch.face_material_model_outputs.densities[q] *
-                     scratch.face_material_model_outputs.specific_heat[q]
-                     :
-                     1.0);
-
-                  Assert (density_c_P >= 0,
-                          ExcMessage ("The product of density and c_P needs to be a "
-                                      "non-negative quantity."));
-
-                  const double conductivity =
-                    ((advection_field.is_temperature())
-                     ?
-                     scratch.face_material_model_outputs.thermal_conductivities[q]
-                     :
-                     0.0);
-                  const double latent_heat_LHS =
-                    ((advection_field.is_temperature())
-                     ?
-                     scratch.face_heating_model_outputs.lhs_latent_heat_terms[q]
-                     :
-                     0.0);
-                  Assert (density_c_P + latent_heat_LHS >= 0,
-                          ExcMessage ("The sum of density times c_P and the latent heat contribution "
-                                      "to the left hand side needs to be a non-negative quantity."));
-
-                  const double penalty = (advection_field.is_temperature()
-                                          ?
-                                          parameters.discontinuous_penalty
-                                          * parameters.temperature_degree
-                                          * parameters.temperature_degree
-                                          / approximate_face_measure(face)
-                                          * conductivity
-                                          / (density_c_P + latent_heat_LHS)
-                                          :
-                                          0.0);
-
-                  const double dirichlet_value = (advection_field.is_temperature()
-                                                  ?
-                                                  this->get_boundary_temperature().boundary_temperature(
-                                                    cell->face(face_no)->boundary_id(),
-                                                    scratch.face_finite_element_values->quadrature_point(q))
-                                                  :
-                                                  this->get_boundary_composition().boundary_composition(
-                                                    cell->face(face_no)->boundary_id(),
-                                                    scratch.face_finite_element_values->quadrature_point(q),
-                                                    advection_field.compositional_variable));
-
-                  Tensor<1,dim> current_u = scratch.face_current_velocity_values[q];
-                  //Subtract off the mesh velocity for ALE corrections if necessary
-                  if (parameters.free_surface_enabled)
-                    current_u -= scratch.face_mesh_velocity_values[q];
-
-                  /**
-                   * The discontinuous Galerkin method uses 2 types of jumps over edges:
-                   * undirected and directed jumps. Undirected jumps are dependent only
-                   * on the order of the numbering of cells. Directed jumps are dependent
-                   * on the direction of the flow. Thus the flow-dependent terms below are
-                   * only calculated if the edge is an inflow edge.
-                   */
-                  const bool inflow = ((current_u * scratch.face_finite_element_values->normal_vector(q)) < 0.);
-
-                  for (unsigned int i=0; i<advection_dofs_per_cell; ++i)
-                    {
-                      data.local_rhs(i)
-                      += (- time_step *  conductivity
-                          * scratch.face_grad_phi_field[i]
-                          * scratch.face_finite_element_values->normal_vector(q)
-                          * dirichlet_value
-
-                          + time_step
-                          * (density_c_P + latent_heat_LHS)
-                          * penalty
-                          * scratch.face_phi_field[i]
-                          * dirichlet_value
-
-                          + (inflow
-                             ?
-                             - (density_c_P + latent_heat_LHS)
-                             * time_step
-                             * (current_u
-                                * scratch.face_finite_element_values->normal_vector(q))
-                             * dirichlet_value
-                             * scratch.face_phi_field[i]
-                             :
-                             0.)
-                         )
-                         *
-                         scratch.face_finite_element_values->JxW(q);
-
-                      // local_matrix terms
-                      for (unsigned int j=0; j<advection_dofs_per_cell; ++j)
-                        {
-                          data.local_matrix(i,j)
-                          += (- time_step *  conductivity
-                              * scratch.face_grad_phi_field[i]
-                              * scratch.face_finite_element_values->normal_vector(q)
-                              * scratch.face_phi_field[j]
-
-                              - time_step *  conductivity
-                              * scratch.face_grad_phi_field[j]
-                              * scratch.face_finite_element_values->normal_vector(q)
-                              * scratch.face_phi_field[i]
-
-                              + time_step
-                              * (density_c_P + latent_heat_LHS)
-                              * penalty
-                              * scratch.face_phi_field[i]
-                              * scratch.face_phi_field[j]
-
-                              + (inflow
-                                 ?
-                                 - (density_c_P + latent_heat_LHS)
-                                 * time_step
-                                 * (current_u
-                                    * scratch.face_finite_element_values->normal_vector(q))
-                                 * scratch.face_phi_field[i]
-                                 * scratch.face_phi_field[j]
-                                 :
-                                 0.)
-                             )
-                             * scratch.face_finite_element_values->JxW(q);
-                        }
-                    }
-                }
-            }
-          else
-            {
-              //Neumann temperature term - no non-zero contribution as only homogeneous Neumann boundary conditions are implemented elsewhere for temperature
-            }
-        }
-
-        void
-        local_assemble_discontinuous_advection_interior_face_terms(const typename DoFHandler<dim>::active_cell_iterator &cell,
-                                                                   const unsigned int face_no,
-                                                                   const typename Simulator<dim>::AdvectionField &advection_field,
-                                                                   internal::Assembly::Scratch::AdvectionSystem<dim> &scratch,
-                                                                   internal::Assembly::CopyData::AdvectionSystem<dim> &data) const
-        {
-          const Parameters<dim> &parameters = this->get_parameters();
-          const Introspection<dim> &introspection = this->introspection();
-          const FiniteElement<dim> &fe = this->get_fe();
-
-          const unsigned int n_q_points    = scratch.face_finite_element_values->n_quadrature_points;
-
-          const double time_step = this->get_timestep();
-
-          if (!advection_field.is_discontinuous(introspection))
-            return;
-
-          // also have the number of dofs that correspond just to the element for
-          // the system we are currently trying to assemble
-          const unsigned int advection_dofs_per_cell = data.local_dof_indices.size();
-          const unsigned int dofs_per_cell = fe.dofs_per_cell;
-
-          Assert (advection_dofs_per_cell < scratch.face_finite_element_values->get_fe().dofs_per_cell, ExcInternalError());
-          Assert (scratch.face_grad_phi_field.size() == advection_dofs_per_cell, ExcInternalError());
-          Assert (scratch.face_phi_field.size() == advection_dofs_per_cell, ExcInternalError());
-
-          const unsigned int solution_component = advection_field.component_index(introspection);
-
-          const FEValuesExtractors::Scalar solution_field = advection_field.scalar_extractor(introspection);
-
-          typename DoFHandler<dim>::face_iterator face = cell->face (face_no);
-
-          //interior face - no contribution on RHS
-          Assert (cell->neighbor(face_no).state() == IteratorState::valid,
-                  ExcInternalError());
-          const typename DoFHandler<dim>::cell_iterator neighbor = cell->neighbor(face_no); //note: NOT active_cell_iterator, so this includes cells that are refined.
-
-          if (!(face->has_children()))
-            {
-              if (!cell->neighbor_is_coarser(face_no) &&
-                  (((neighbor->is_locally_owned()) && (cell->index() < neighbor->index()))
-                   ||
-                   ((!neighbor->is_locally_owned()) && (cell->subdomain_id() < neighbor->subdomain_id()))))
-                {
-                  Assert (cell->is_locally_owned(), ExcInternalError());
-                  //cell and neighbor are equal-sized, and cell has been chosen to assemble this face, so calculate from cell
-
-                  //how does the neighbor talk about this cell?
-                  const unsigned int neighbor2=cell->neighbor_of_neighbor(face_no);
-
-                  //set up neighbor values
-                  scratch.neighbor_face_finite_element_values->reinit (neighbor, neighbor2);
-
-                  this->compute_material_model_input_values (this->get_current_linearization_point(),
-                                                             *scratch.neighbor_face_finite_element_values,
-                                                             neighbor,
-                                                             true,
-                                                             scratch.neighbor_face_material_model_inputs);
-                  this->get_material_model().evaluate(scratch.neighbor_face_material_model_inputs,
-                                                      scratch.neighbor_face_material_model_outputs);
-
-                  this->get_heating_model_manager().evaluate(scratch.neighbor_face_material_model_inputs,
-                                                             scratch.neighbor_face_material_model_outputs,
-                                                             scratch.neighbor_face_heating_model_outputs);
-
-                  std::vector<types::global_dof_index> neighbor_dof_indices (dofs_per_cell);
-                  // get all dof indices on the neighbor, then extract those
-                  // that correspond to the solution_field we are interested in
-                  neighbor->get_dof_indices (neighbor_dof_indices);
-                  for (unsigned int i=0, i_advection=0; i_advection<advection_dofs_per_cell;/*increment at end of loop*/)
-                    {
-                      if (fe.system_to_component_index(i).first == solution_component)
-                        {
-                          data.neighbor_dof_indices[face_no * GeometryInfo<dim>::max_children_per_face][i_advection] = neighbor_dof_indices[i];
-                          ++i_advection;
-                        }
-                      ++i;
-                    }
-                  data.assembled_matrices[face_no * GeometryInfo<dim>::max_children_per_face] = true;
-
-                  for (unsigned int q=0; q<n_q_points; ++q)
-                    {
-                      // precompute the values of shape functions and their gradients.
-                      // We only need to look up values of shape functions if they
-                      // belong to 'our' component. They are zero otherwise anyway.
-                      // Note that we later only look at the values that we do set here.
-                      for (unsigned int i=0, i_advection=0; i_advection<advection_dofs_per_cell;/*increment at end of loop*/)
-                        {
-                          if (fe.system_to_component_index(i).first == solution_component)
-                            {
-                              scratch.face_grad_phi_field[i_advection]          = (*scratch.face_finite_element_values)[solution_field].gradient (i, q);
-                              scratch.face_phi_field[i_advection]               = (*scratch.face_finite_element_values)[solution_field].value (i, q);
-                              scratch.neighbor_face_grad_phi_field[i_advection] = (*scratch.neighbor_face_finite_element_values)[solution_field].gradient (i, q);
-                              scratch.neighbor_face_phi_field[i_advection]      = (*scratch.neighbor_face_finite_element_values)[solution_field].value (i, q);
-                              ++i_advection;
-                            }
-                          ++i;
-                        }
-
-                      const double density_c_P              =
-                        ((advection_field.is_temperature())
-                         ?
-                         scratch.face_material_model_outputs.densities[q] *
-                         scratch.face_material_model_outputs.specific_heat[q]
-                         :
-                         1.0);
-
-                      Assert (density_c_P >= 0,
-                              ExcMessage ("The product of density and c_P needs to be a "
-                                          "non-negative quantity."));
-
-                      const double conductivity =
-                        ((advection_field.is_temperature())
-                         ?
-                         scratch.face_material_model_outputs.thermal_conductivities[q]
-                         :
-                         0.0);
-                      const double latent_heat_LHS =
-                        ((advection_field.is_temperature())
-                         ?
-                         scratch.face_heating_model_outputs.lhs_latent_heat_terms[q]
-                         :
-                         0.0);
-                      Assert (density_c_P + latent_heat_LHS >= 0,
-                              ExcMessage ("The sum of density times c_P and the latent heat contribution "
-                                          "to the left hand side needs to be a non-negative quantity."));
-
-                      const double penalty = (advection_field.is_temperature()
-                                              ?
-                                              parameters.discontinuous_penalty
-                                              * parameters.temperature_degree
-                                              * parameters.temperature_degree
-                                              / approximate_face_measure(face)
-                                              * conductivity
-                                              / (density_c_P + latent_heat_LHS)
-                                              :
-                                              0.0);
-
-                      Tensor<1,dim> current_u = scratch.face_current_velocity_values[q];
-                      //Subtract off the mesh velocity for ALE corrections if necessary
-                      if (parameters.free_surface_enabled)
-                        current_u -= scratch.face_mesh_velocity_values[q];
-
-                      const double neighbor_density_c_P              =
-                        ((advection_field.is_temperature())
-                         ?
-                         scratch.neighbor_face_material_model_outputs.densities[q] *
-                         scratch.neighbor_face_material_model_outputs.specific_heat[q]
-                         :
-                         1.0);
-
-                      Assert (neighbor_density_c_P >= 0,
-                              ExcMessage ("The product of density and c_P on the neighbor needs to be a "
-                                          "non-negative quantity."));
-
-                      const double neighbor_conductivity =
-                        ((advection_field.is_temperature())
-                         ?
-                         scratch.neighbor_face_material_model_outputs.thermal_conductivities[q]
-                         :
-                         0.0);
-                      const double neighbor_latent_heat_LHS =
-                        ((advection_field.is_temperature())
-                         ?
-                         scratch.neighbor_face_heating_model_outputs.lhs_latent_heat_terms[q]
-                         :
-                         0.0);
-                      Assert (neighbor_density_c_P + neighbor_latent_heat_LHS >= 0,
-                              ExcMessage ("The sum of density times c_P and the latent heat contribution "
-                                          "to the left hand side on the neighbor needs to be a non-negative quantity."));
-
-                      const double neighbor_penalty = (advection_field.is_temperature()
-                                                       ?
-                                                       parameters.discontinuous_penalty
-                                                       * parameters.temperature_degree
-                                                       * parameters.temperature_degree
-                                                       / approximate_face_measure(neighbor->face(neighbor2))
-                                                       * neighbor_conductivity
-                                                       / (neighbor_density_c_P + neighbor_latent_heat_LHS)
-                                                       :
-                                                       0.0);
-
-                      const double max_penalty = std::max(penalty, neighbor_penalty);
-
-                      const double max_density_c_P_and_latent_heat =
-                        std::max(density_c_P + latent_heat_LHS,
-                                 neighbor_density_c_P + neighbor_latent_heat_LHS);
-
-                      Assert (numbers::is_finite(max_density_c_P_and_latent_heat),
-                              ExcMessage ("The maximum product of density and c_P plus latent heat LHS on the neighbor needs to be a finite quantity."));
-                      Assert (max_density_c_P_and_latent_heat >= 0,
-                              ExcMessage ("The maximum product of density and c_P plus latent heat LHS on the neighbor needs to be a "
-                                          "non-negative quantity."));
-
-                      /**
-                       * The discontinuous Galerkin method uses 2 types of jumps over edges:
-                       * undirected and directed jumps. Undirected jumps are dependent only
-                       * on the order of the numbering of cells. Directed jumps are dependent
-                       * on the direction of the flow. Thus the flow-dependent terms below are
-                       * only calculated if the edge is an inflow edge.
-                       */
-                      const bool inflow = ((current_u * scratch.face_finite_element_values->normal_vector(q)) < 0.);
-
-                      for (unsigned int i=0; i<advection_dofs_per_cell; ++i)
-                        {
-                          for (unsigned int j=0; j<advection_dofs_per_cell; ++j)
-                            {
-                              data.local_matrix(i,j)
-                              += (- 0.5 * time_step * conductivity
-                                  * scratch.face_grad_phi_field[i]
-                                  * scratch.face_finite_element_values->normal_vector(q)
-                                  * scratch.face_phi_field[j]
-
-                                  - 0.5 * time_step * conductivity
-                                  * scratch.face_grad_phi_field[j]
-                                  * scratch.face_finite_element_values->normal_vector(q)
-                                  * scratch.face_phi_field[i]
-
-                                  + time_step
-                                  * max_density_c_P_and_latent_heat
-                                  * max_penalty
-                                  * scratch.face_phi_field[i]
-                                  * scratch.face_phi_field[j]
-
-                                  - (inflow
-                                     ?
-                                     (density_c_P + latent_heat_LHS)
-                                     * time_step
-                                     * (current_u
-                                        * scratch.face_finite_element_values->normal_vector(q))
-                                     * scratch.face_phi_field[i]
-                                     * scratch.face_phi_field[j]
-                                     :
-                                     0.)
-                                 )
-                                 * scratch.face_finite_element_values->JxW(q);
-
-                              data.local_matrices_int_ext[face_no * GeometryInfo<dim>::max_children_per_face](i,j)
-                              += (- 0.5 * time_step * neighbor_conductivity
-                                  * scratch.neighbor_face_grad_phi_field[j]
-                                  * scratch.face_finite_element_values->normal_vector(q)
-                                  * scratch.face_phi_field[i]
-
-                                  + 0.5 * time_step * conductivity
-                                  * scratch.face_grad_phi_field[i]
-                                  * scratch.face_finite_element_values->normal_vector(q)
-                                  * scratch.neighbor_face_phi_field[j]
-
-                                  - time_step
-                                  * max_density_c_P_and_latent_heat
-                                  * max_penalty
-                                  * scratch.neighbor_face_phi_field[j]
-                                  * scratch.face_phi_field[i]
-
-                                  + (inflow
-                                     ?
-                                     (density_c_P + latent_heat_LHS)
-                                     * time_step
-                                     * (current_u
-                                        * scratch.face_finite_element_values->normal_vector(q))
-                                     * scratch.face_phi_field[i]
-                                     * scratch.neighbor_face_phi_field[j]
-                                     :
-                                     0.)
-                                 )
-                                 * scratch.face_finite_element_values->JxW(q);
-
-                              data.local_matrices_ext_int[face_no * GeometryInfo<dim>::max_children_per_face](i,j)
-                              += (+ 0.5 * time_step * conductivity
-                                  * scratch.face_grad_phi_field[j]
-                                  * scratch.face_finite_element_values->normal_vector(q)
-                                  * scratch.neighbor_face_phi_field[i]
-
-                                  - 0.5 * time_step * neighbor_conductivity
-                                  * scratch.neighbor_face_grad_phi_field[i]
-                                  * scratch.face_finite_element_values->normal_vector(q)
-                                  * scratch.face_phi_field[j]
-
-                                  - time_step
-                                  * max_density_c_P_and_latent_heat
-                                  * max_penalty
-                                  * scratch.face_phi_field[j]
-                                  * scratch.neighbor_face_phi_field[i]
-
-                                  - (!inflow
-                                     ?
-                                     (neighbor_density_c_P + neighbor_latent_heat_LHS)
-                                     * time_step
-                                     * (current_u
-                                        * scratch.face_finite_element_values->normal_vector(q))
-                                     * scratch.neighbor_face_phi_field[i]
-                                     * scratch.face_phi_field[j]
-                                     :
-                                     0.)
-                                 )
-                                 * scratch.face_finite_element_values->JxW(q);
-
-                              data.local_matrices_ext_ext[face_no * GeometryInfo<dim>::max_children_per_face](i,j)
-                              += (+ 0.5 * time_step * neighbor_conductivity
-                                  * scratch.neighbor_face_grad_phi_field[i]
-                                  * scratch.face_finite_element_values->normal_vector(q)
-                                  * scratch.neighbor_face_phi_field[j]
-
-                                  + 0.5 * time_step * neighbor_conductivity
-                                  * scratch.neighbor_face_grad_phi_field[j]
-                                  * scratch.face_finite_element_values->normal_vector(q)
-                                  * scratch.neighbor_face_phi_field[i]
-
-                                  + time_step
-                                  * max_density_c_P_and_latent_heat
-                                  * max_penalty
-                                  * scratch.neighbor_face_phi_field[i]
-                                  * scratch.neighbor_face_phi_field[j]
-
-                                  + (!inflow
-                                     ?
-                                     (neighbor_density_c_P + neighbor_latent_heat_LHS)
-                                     * time_step
-                                     * (current_u
-                                        * scratch.face_finite_element_values->normal_vector(q))
-                                     * scratch.neighbor_face_phi_field[i]
-                                     * scratch.neighbor_face_phi_field[j]
-                                     :
-                                     0.)
-                                 )
-                                 * scratch.face_finite_element_values->JxW(q);
-                            }
-                        }
-                    }
-                }
-              else
-                {
-                  /* neighbor is taking responsibility for assembly of this face, because
-                   * either (1) neighbor is coarser, or
-                   *        (2) neighbor is equally-sized and
-                   *           (a) neighbor is on a different subdomain, with lower subdmain_id(), or
-                   *           (b) neighbor is on the same subdomain and has lower index().
-                  */
-                }
-            }
-          else //face->has_children(), so always assemble from here.
-            {
-              //how does the neighbor talk about this cell?
-              const unsigned int neighbor2 = cell->neighbor_face_no(face_no);
-
-              //loop over subfaces
-              for (unsigned int subface_no=0; subface_no<face->number_of_children(); ++subface_no)
-                {
-                  const typename DoFHandler<dim>::active_cell_iterator neighbor_child
-                    = cell->neighbor_child_on_subface (face_no, subface_no);
-
-                  //set up subface values
-                  scratch.subface_finite_element_values->reinit (cell, face_no, subface_no);
-
-                  //subface->face
-                  (*scratch.subface_finite_element_values)[introspection.extractors.velocities].get_function_values(this->get_current_linearization_point(),
-                      scratch.face_current_velocity_values);
-
-                  //get the mesh velocity, as we need to subtract it off of the advection systems
-                  if (parameters.free_surface_enabled)
-                    (*scratch.subface_finite_element_values)[introspection.extractors.velocities].get_function_values(this->get_mesh_velocity(),
-                        scratch.face_mesh_velocity_values);
-
-                  //get the mesh velocity, as we need to subtract it off of the advection systems
-                  if (parameters.free_surface_enabled)
-                    (*scratch.subface_finite_element_values)[introspection.extractors.velocities].get_function_values(this->get_mesh_velocity(),
-                        scratch.face_mesh_velocity_values);
-
-                  this->compute_material_model_input_values (this->get_current_linearization_point(),
-                                                             *scratch.subface_finite_element_values,
-                                                             cell,
-                                                             true,
-                                                             scratch.face_material_model_inputs);
-                  this->get_material_model().evaluate(scratch.face_material_model_inputs,
-                                                      scratch.face_material_model_outputs);
-
-                  this->get_heating_model_manager().evaluate(scratch.face_material_model_inputs,
-                                                             scratch.face_material_model_outputs,
-                                                             scratch.face_heating_model_outputs);
-
-                  //set up neighbor values
-                  scratch.neighbor_face_finite_element_values->reinit (neighbor_child, neighbor2);
-
-                  this->compute_material_model_input_values (this->get_current_linearization_point(),
-                                                             *scratch.neighbor_face_finite_element_values,
-                                                             neighbor_child,
-                                                             true,
-                                                             scratch.neighbor_face_material_model_inputs);
-                  this->get_material_model().evaluate(scratch.neighbor_face_material_model_inputs,
-                                                      scratch.neighbor_face_material_model_outputs);
-
-                  this->get_heating_model_manager().evaluate(scratch.neighbor_face_material_model_inputs,
-                                                             scratch.neighbor_face_material_model_outputs,
-                                                             scratch.neighbor_face_heating_model_outputs);
-
-                  std::vector<types::global_dof_index> neighbor_dof_indices (fe.dofs_per_cell);
-                  // get all dof indices on the neighbor, then extract those
-                  // that correspond to the solution_field we are interested in
-                  neighbor_child->get_dof_indices (neighbor_dof_indices);
-                  for (unsigned int i=0, i_advection=0; i_advection<advection_dofs_per_cell;/*increment at end of loop*/)
-                    {
-                      if (fe.system_to_component_index(i).first == solution_component)
-                        {
-                          data.neighbor_dof_indices[face_no * GeometryInfo<dim>::max_children_per_face + subface_no][i_advection] = neighbor_dof_indices[i];
-                          ++i_advection;
-                        }
-                      ++i;
-                    }
-                  data.assembled_matrices[face_no * GeometryInfo<dim>::max_children_per_face + subface_no] = true;
-
-                  for (unsigned int q=0; q<n_q_points; ++q)
-                    {
-                      // precompute the values of shape functions and their gradients.
-                      // We only need to look up values of shape functions if they
-                      // belong to 'our' component. They are zero otherwise anyway.
-                      // Note that we later only look at the values that we do set here.
-                      for (unsigned int i=0, i_advection=0; i_advection<advection_dofs_per_cell; /*increment at end of loop*/)
-                        {
-                          if (fe.system_to_component_index(i).first == solution_component)
-                            {
-                              scratch.face_grad_phi_field[i_advection]          = (*scratch.subface_finite_element_values)[solution_field].gradient (i, q);
-                              scratch.face_phi_field[i_advection]               = (*scratch.subface_finite_element_values)[solution_field].value (i, q);
-                              scratch.neighbor_face_grad_phi_field[i_advection] = (*scratch.neighbor_face_finite_element_values)[solution_field].gradient (i, q);
-                              scratch.neighbor_face_phi_field[i_advection]      = (*scratch.neighbor_face_finite_element_values)[solution_field].value (i, q);
-                              ++i_advection;
-                            }
-                          ++i;
-                        }
-
-                      const double density_c_P              =
-                        ((advection_field.is_temperature())
-                         ?
-                         scratch.face_material_model_outputs.densities[q] *
-                         scratch.face_material_model_outputs.specific_heat[q]
-                         :
-                         1.0);
-
-                      Assert (density_c_P >= 0,
-                              ExcMessage ("The product of density and c_P needs to be a "
-                                          "non-negative quantity."));
-
-                      const double conductivity =
-                        ((advection_field.is_temperature())
-                         ?
-                         scratch.face_material_model_outputs.thermal_conductivities[q]
-                         :
-                         0.0);
-                      const double latent_heat_LHS =
-                        ((advection_field.is_temperature())
-                         ?
-                         scratch.face_heating_model_outputs.lhs_latent_heat_terms[q]
-                         :
-                         0.0);
-                      Assert (density_c_P + latent_heat_LHS >= 0,
-                              ExcMessage ("The sum of density times c_P and the latent heat contribution "
-                                          "to the left hand side needs to be a non-negative quantity."));
-
-                      const double penalty = (advection_field.is_temperature()
-                                              ?
-                                              parameters.discontinuous_penalty
-                                              * parameters.temperature_degree
-                                              * parameters.temperature_degree
-                                              / approximate_face_measure(face)
-                                              * conductivity
-                                              / (density_c_P + latent_heat_LHS)
-                                              :
-                                              0.0);
-
-                      Tensor<1,dim> current_u = scratch.face_current_velocity_values[q];
-                      //Subtract off the mesh velocity for ALE corrections if necessary
-                      if (parameters.free_surface_enabled)
-                        current_u -= scratch.face_mesh_velocity_values[q];
-
-                      const double neighbor_density_c_P              =
-                        ((advection_field.is_temperature())
-                         ?
-                         scratch.neighbor_face_material_model_outputs.densities[q] *
-                         scratch.neighbor_face_material_model_outputs.specific_heat[q]
-                         :
-                         1.0);
-
-                      Assert (neighbor_density_c_P >= 0,
-                              ExcMessage ("The product of density and c_P on the neighbor needs to be a "
-                                          "non-negative quantity."));
-
-                      const double neighbor_conductivity =
-                        ((advection_field.is_temperature())
-                         ?
-                         scratch.neighbor_face_material_model_outputs.thermal_conductivities[q]
-                         :
-                         0.0);
-                      const double neighbor_latent_heat_LHS =
-                        ((advection_field.is_temperature())
-                         ?
-                         scratch.neighbor_face_heating_model_outputs.lhs_latent_heat_terms[q]
-                         :
-                         0.0);
-                      Assert (neighbor_density_c_P + neighbor_latent_heat_LHS >= 0,
-                              ExcMessage ("The sum of density times c_P and the latent heat contribution "
-                                          "to the left hand side on the neighbor needs to be a non-negative quantity."));
-
-                      const double neighbor_penalty = (advection_field.is_temperature()
-                                                       ?
-                                                       parameters.discontinuous_penalty
-                                                       * parameters.temperature_degree
-                                                       * parameters.temperature_degree
-                                                       / approximate_face_measure(neighbor_child->face(neighbor2))
-                                                       * neighbor_conductivity
-                                                       / (neighbor_density_c_P + neighbor_latent_heat_LHS)
-                                                       :
-                                                       0.0);
-
-                      const double max_penalty = std::max(penalty, neighbor_penalty);
-
-                      const double max_density_c_P_and_latent_heat =
-                        std::max(density_c_P + latent_heat_LHS,
-                                 neighbor_density_c_P + neighbor_latent_heat_LHS);
-
-                      Assert (numbers::is_finite(max_density_c_P_and_latent_heat),
-                              ExcMessage ("The maximum product of density and c_P plus latent heat LHS on the neighbor needs to be a finite quantity."));
-                      Assert (max_density_c_P_and_latent_heat >= 0,
-                              ExcMessage ("The maximum product of density and c_P plus latent heat LHS on the neighbor needs to be a "
-                                          "non-negative quantity."));
-
-                      /**
-                       * The discontinuous Galerkin method uses 2 types of jumps over edges:
-                       * undirected and directed jumps. Undirected jumps are dependent only
-                       * on the order of the numbering of cells. Directed jumps are dependent
-                       * on the direction of the flow. Thus the flow-dependent terms below are
-                       * only calculated if the edge is an inflow edge.
-                       */
-                      const bool inflow = ((current_u * scratch.subface_finite_element_values->normal_vector(q)) < 0.);
-
-                      for (unsigned int i=0; i<advection_dofs_per_cell; ++i)
-                        {
-                          for (unsigned int j=0; j<advection_dofs_per_cell; ++j)
-                            {
-                              data.local_matrix(i,j)
-                              += (- 0.5 * time_step * conductivity
-                                  * scratch.face_grad_phi_field[i]
-                                  * scratch.subface_finite_element_values->normal_vector(q)
-                                  * scratch.face_phi_field[j]
-
-                                  - 0.5 * time_step * conductivity
-                                  * scratch.face_grad_phi_field[j]
-                                  * scratch.subface_finite_element_values->normal_vector(q)
-                                  * scratch.face_phi_field[i]
-
-                                  + time_step
-                                  * max_density_c_P_and_latent_heat
-                                  * max_penalty
-                                  * scratch.face_phi_field[i]
-                                  * scratch.face_phi_field[j]
-
-                                  - (inflow
-                                     ?
-                                     (density_c_P + latent_heat_LHS)
-                                     * time_step
-                                     * (current_u
-                                        * scratch.subface_finite_element_values->normal_vector(q))
-                                     * scratch.face_phi_field[i]
-                                     * scratch.face_phi_field[j]
-                                     :
-                                     0.)
-                                 )
-                                 * scratch.subface_finite_element_values->JxW(q);
-
-                              data.local_matrices_int_ext[face_no * GeometryInfo<dim>::max_children_per_face + subface_no](i,j)
-                              += (- 0.5 * time_step * neighbor_conductivity
-                                  * scratch.neighbor_face_grad_phi_field[j]
-                                  * scratch.subface_finite_element_values->normal_vector(q)
-                                  * scratch.face_phi_field[i]
-
-                                  + 0.5 * time_step * conductivity
-                                  * scratch.face_grad_phi_field[i]
-                                  * scratch.subface_finite_element_values->normal_vector(q)
-                                  * scratch.neighbor_face_phi_field[j]
-
-                                  - time_step
-                                  * max_density_c_P_and_latent_heat
-                                  * max_penalty
-                                  * scratch.neighbor_face_phi_field[j]
-                                  * scratch.face_phi_field[i]
-
-                                  + (inflow
-                                     ?
-                                     (density_c_P + latent_heat_LHS)
-                                     * time_step
-                                     * (current_u
-                                        * scratch.subface_finite_element_values->normal_vector(q))
-                                     * scratch.face_phi_field[i]
-                                     * scratch.neighbor_face_phi_field[j]
-                                     :
-                                     0.)
-                                 )
-                                 * scratch.subface_finite_element_values->JxW(q);
-
-                              data.local_matrices_ext_int[face_no * GeometryInfo<dim>::max_children_per_face + subface_no](i,j)
-                              += (+ 0.5 * time_step * conductivity
-                                  * scratch.face_grad_phi_field[j]
-                                  * scratch.subface_finite_element_values->normal_vector(q)
-                                  * scratch.neighbor_face_phi_field[i]
-
-                                  - 0.5 * time_step * neighbor_conductivity
-                                  * scratch.neighbor_face_grad_phi_field[i]
-                                  * scratch.subface_finite_element_values->normal_vector(q)
-                                  * scratch.face_phi_field[j]
-
-                                  - time_step
-                                  * max_density_c_P_and_latent_heat
-                                  * max_penalty
-                                  * scratch.face_phi_field[j]
-                                  * scratch.neighbor_face_phi_field[i]
-
-                                  - (!inflow
-                                     ?
-                                     (neighbor_density_c_P + neighbor_latent_heat_LHS)
-                                     * time_step
-                                     * (current_u
-                                        * scratch.subface_finite_element_values->normal_vector(q))
-                                     * scratch.neighbor_face_phi_field[i]
-                                     * scratch.face_phi_field[j]
-                                     :
-                                     0.)
-                                 )
-                                 * scratch.subface_finite_element_values->JxW(q);
-
-                              data.local_matrices_ext_ext[face_no * GeometryInfo<dim>::max_children_per_face + subface_no](i,j)
-                              += (+ 0.5 * time_step * neighbor_conductivity
-                                  * scratch.neighbor_face_grad_phi_field[i]
-                                  * scratch.subface_finite_element_values->normal_vector(q)
-                                  * scratch.neighbor_face_phi_field[j]
-
-                                  + 0.5 * time_step * neighbor_conductivity
-                                  * scratch.neighbor_face_grad_phi_field[j]
-                                  * scratch.subface_finite_element_values->normal_vector(q)
-                                  * scratch.neighbor_face_phi_field[i]
-
-                                  + time_step
-                                  * max_density_c_P_and_latent_heat
-                                  * max_penalty
-                                  * scratch.neighbor_face_phi_field[i]
-                                  * scratch.neighbor_face_phi_field[j]
-
-                                  + (!inflow
-                                     ?
-                                     (neighbor_density_c_P + neighbor_latent_heat_LHS)
-                                     * time_step
-                                     * (current_u
-                                        * scratch.subface_finite_element_values->normal_vector(q))
-                                     * scratch.neighbor_face_phi_field[i]
-                                     * scratch.neighbor_face_phi_field[j]
-                                     :
-                                     0.)
-                                 )
-                                 * scratch.subface_finite_element_values->JxW(q);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    };
-
-
     /**
      * A namespace for the definition of functions that implement various
      * other terms that need to occasionally or always be assembled.
@@ -2322,8 +603,10 @@ namespace aspect
     // create an object for the complete equations assembly; add its
     // member functions to the signals and add the object the list
     // of assembler objects
-    aspect::Assemblers::CompleteEquations<dim> *complete_equation_assembler
-      = new aspect::Assemblers::CompleteEquations<dim>();
+    aspect::StokesAssembler<dim> *stokes_assembler
+      = new aspect::StokesAssembler<dim>();
+    aspect::AdvectionAssembler<dim> *adv_assembler
+      = new aspect::AdvectionAssembler<dim>();
 
     aspect::Assemblers::MeltEquations<dim> *melt_equation_assembler = NULL;
     if (parameters.include_melt_transport)
@@ -2336,8 +619,8 @@ namespace aspect
                                 std_cxx11::_1, std_cxx11::_2, std_cxx11::_3));
     else
       assemblers->local_assemble_stokes_preconditioner
-      .connect (std_cxx11::bind(&aspect::Assemblers::CompleteEquations<dim>::local_assemble_stokes_preconditioner,
-                                std_cxx11::cref (*complete_equation_assembler),
+      .connect (std_cxx11::bind(&aspect::StokesAssembler<dim>::local_assemble_stokes_preconditioner,
+                                std_cxx11::cref (*stokes_assembler),
                                 std_cxx11::_1, std_cxx11::_2, std_cxx11::_3));
 
     if (parameters.include_melt_transport)
@@ -2349,27 +632,75 @@ namespace aspect
                                 std_cxx11::_3,
                                 std_cxx11::_4,
                                 std_cxx11::_5));
-    else if (material_model->is_compressible())
-      assemblers->local_assemble_stokes_system
-      .connect (std_cxx11::bind(&aspect::Assemblers::CompleteEquations<dim>::local_assemble_stokes_system_compressible,
-                                std_cxx11::cref (*complete_equation_assembler),
-                                // discard cell,
-                                std_cxx11::_2,
-                                std_cxx11::_3,
-                                std_cxx11::_4,
-                                std_cxx11::_5));
     else
-      assemblers->local_assemble_stokes_system
-      .connect (std_cxx11::bind(&aspect::Assemblers::CompleteEquations<dim>::local_assemble_stokes_system_incompressible,
-                                std_cxx11::cref (*complete_equation_assembler),
-                                // discard cell,
-                                std_cxx11::_2,
-                                std_cxx11::_3,
-                                std_cxx11::_4,
-                                std_cxx11::_5));
+      {
+        assemblers->local_assemble_stokes_system
+        .connect (std_cxx11::bind(&aspect::StokesAssembler<dim>::local_assemble_stokes_incompressible,
+                                  std_cxx11::cref (*stokes_assembler),
+                                  // discard cell,
+                                  std_cxx11::_2,
+                                  std_cxx11::_3,
+                                  std_cxx11::_4,
+                                  std_cxx11::_5));
+
+        if (material_model->is_compressible())
+          assemblers->local_assemble_stokes_system
+          .connect (std_cxx11::bind(&aspect::StokesAssembler<dim>::local_assemble_stokes_compressible_diffusion,
+                                    std_cxx11::cref (*stokes_assembler),
+                                    // discard cell,
+                                    std_cxx11::_2,
+                                    std_cxx11::_3,
+                                    std_cxx11::_4,
+                                    std_cxx11::_5));
+
+        if (parameters.formulation_mass == Parameters<dim>::FormulationType::implicit_adiabatic)
+          assemblers->local_assemble_stokes_system
+          .connect (std_cxx11::bind(&aspect::StokesAssembler<dim>::local_assemble_stokes_mass_density_implicit,
+                                    std_cxx11::cref (*stokes_assembler),
+                                    // discard cell,
+                                    std_cxx11::_2,
+                                    std_cxx11::_3,
+                                    std_cxx11::_4,
+                                    std_cxx11::_5,
+                                    std_cxx11::cref (this->parameters)));
+        else if (parameters.formulation_mass == Parameters<dim>::FormulationType::adiabatic)
+          {
+            assemblers->local_assemble_stokes_system
+            .connect (std_cxx11::bind(&aspect::StokesAssembler<dim>::local_assemble_stokes_mass_density_gradient,
+                                      std_cxx11::cref (*stokes_assembler),
+                                      // discard cell,
+                                      std_cxx11::_2,
+                                      std_cxx11::_3,
+                                      std_cxx11::_4,
+                                      std_cxx11::_5,
+                                      std_cxx11::cref (this->parameters)));
+          }
+        else if (parameters.formulation_mass == Parameters<dim>::FormulationType::incompressible
+                 || (parameters.formulation_mass == Parameters<dim>::FormulationType::ask_material_model
+                     && !material_model->is_compressible()))
+          {
+            // do nothing, because we assembled div u =0 above already
+          }
+        else
+          assemblers->local_assemble_stokes_system
+          .connect (std_cxx11::bind(&aspect::StokesAssembler<dim>::local_assemble_stokes_mass_density_explicit,
+                                    std_cxx11::cref (*stokes_assembler),
+                                    // discard cell,
+                                    std_cxx11::_2,
+                                    std_cxx11::_3,
+                                    std_cxx11::_4,
+                                    std_cxx11::_5,
+                                    std_cxx11::cref (this->parameters)));
+
+      }
+
 
     assembler_objects.push_back (std_cxx11::shared_ptr<internal::Assembly::Assemblers::AssemblerBase<dim> >
-                                 (complete_equation_assembler));
+                                 (stokes_assembler));
+
+    assembler_objects.push_back (std_cxx11::shared_ptr<internal::Assembly::Assemblers::AssemblerBase<dim> >
+                                 (adv_assembler));
+
 
     if (parameters.include_melt_transport)
       assembler_objects.push_back (std_cxx11::shared_ptr<internal::Assembly::Assemblers::AssemblerBase<dim> >
@@ -2448,8 +779,8 @@ namespace aspect
     else
       {
         assemblers->local_assemble_advection_system
-        .connect (std_cxx11::bind(&aspect::Assemblers::CompleteEquations<dim>::local_assemble_advection_system,
-                                  std_cxx11::cref (*complete_equation_assembler),
+        .connect (std_cxx11::bind(&aspect::AdvectionAssembler<dim>::local_assemble_advection_system,
+                                  std_cxx11::cref (*adv_assembler),
                                   // discard cell,
                                   std_cxx11::_2,
                                   std_cxx11::_3,
@@ -2457,19 +788,20 @@ namespace aspect
                                   std_cxx11::_5));
 
         assemblers->compute_advection_system_residual
-        .connect (std_cxx11::bind(&aspect::Assemblers::CompleteEquations<dim>::compute_advection_system_residual,
-                                  std_cxx11::cref (*complete_equation_assembler),
+        .connect (std_cxx11::bind(&aspect::AdvectionAssembler<dim>::compute_advection_system_residual,
+                                  std_cxx11::cref (*adv_assembler),
                                   // discard cell,
                                   std_cxx11::_2,
                                   std_cxx11::_3));
+
       }
 
     if (parameters.use_discontinuous_temperature_discretization ||
         parameters.use_discontinuous_composition_discretization)
       {
         assemblers->local_assemble_advection_system_on_interior_face
-        .connect(std_cxx11::bind(&aspect::Assemblers::CompleteEquations<dim>::local_assemble_discontinuous_advection_interior_face_terms,
-                                 std_cxx11::cref (*complete_equation_assembler),
+        .connect(std_cxx11::bind(&aspect::AdvectionAssembler<dim>::local_assemble_discontinuous_advection_interior_face_terms,
+                                 std_cxx11::cref (*adv_assembler),
                                  std_cxx11::_1,
                                  std_cxx11::_2,
                                  std_cxx11::_3,
@@ -2477,8 +809,8 @@ namespace aspect
                                  std_cxx11::_5));
 
         assemblers->local_assemble_advection_system_on_boundary_face
-        .connect(std_cxx11::bind(&aspect::Assemblers::CompleteEquations<dim>::local_assemble_discontinuous_advection_boundary_face_terms,
-                                 std_cxx11::cref (*complete_equation_assembler),
+        .connect(std_cxx11::bind(&aspect::AdvectionAssembler<dim>::local_assemble_discontinuous_advection_boundary_face_terms,
+                                 std_cxx11::cref (*adv_assembler),
                                  std_cxx11::_1,
                                  std_cxx11::_2,
                                  std_cxx11::_3,
@@ -2748,6 +1080,33 @@ namespace aspect
     scratch.finite_element_values[introspection.extractors.velocities].get_function_values(current_linearization_point,
         scratch.velocity_values);
 
+    if (parameters.formulation_mass == Parameters<dim>::FormulationType::adiabatic_density
+        || parameters.formulation_mass == Parameters<dim>::FormulationType::implicit_adiabatic
+        || parameters.formulation_mass == Parameters<dim>::FormulationType::adiabatic
+       )
+      {
+        const unsigned int n_q_points = scratch.finite_element_values.n_quadrature_points;
+        scratch.mass_densities.resize(n_q_points);
+        for (unsigned int q=0; q<n_q_points; ++q)
+          {
+            scratch.mass_densities[q] = adiabatic_conditions->density(scratch.material_model_inputs.position[q]);
+          }
+      }
+    else
+      scratch.mass_densities = scratch.material_model_outputs.densities;
+
+    if (parameters.formulation_mass == Parameters<dim>::FormulationType::adiabatic)
+      {
+        const unsigned int n_q_points = scratch.finite_element_values.n_quadrature_points;
+        scratch.adiabatic_density_gradients.resize(n_q_points);
+        for (unsigned int q=0; q<n_q_points; ++q)
+          {
+            scratch.adiabatic_density_gradients[q] = adiabatic_conditions->density_derivative(scratch.material_model_inputs.position[q]);
+          }
+      }
+    else
+      scratch.adiabatic_density_gradients.resize(0);
+
     // trigger the invocation of the various functions that actually do
     // all of the assembling
     assemblers->local_assemble_stokes_system(cell, pressure_scaling, rebuild_stokes_matrix,
@@ -2771,11 +1130,13 @@ namespace aspect
 
               material_model->evaluate(scratch.face_material_model_inputs,
                                        scratch.face_material_model_outputs);
-//TODO: the following doesn't currently compile because the get_quadrature() call returns
+
+              // TODO: formulations?
+
 //  a dim-1 dimensional quadrature
               // MaterialModel::MaterialAveraging::average (parameters.material_averaging,
               //                                            cell,
-              //                                            scratch.face_finite_element_values.get_quadrature(),
+//                                     compressibility * density *
               //                                            scratch.face_finite_element_values.get_mapping(),
               //                                            scratch.face_material_model_outputs);
             }
@@ -3010,6 +1371,15 @@ namespace aspect
 
     material_model->evaluate(scratch.material_model_inputs,
                              scratch.material_model_outputs);
+    if (parameters.formulation_temperature == Parameters<dim>::FormulationType::adiabatic)
+      {
+        const unsigned int n_q_points = scratch.finite_element_values.n_quadrature_points;
+        for (unsigned int q=0; q<n_q_points; ++q)
+          {
+            scratch.material_model_outputs.densities[q] = adiabatic_conditions->density(scratch.material_model_inputs.position[q]);
+          }
+      }
+
     MaterialModel::MaterialAveraging::average (parameters.material_averaging,
                                                cell,
                                                scratch.finite_element_values.get_quadrature(),
@@ -3019,7 +1389,6 @@ namespace aspect
     heating_model_manager.evaluate(scratch.material_model_inputs,
                                    scratch.material_model_outputs,
                                    scratch.heating_model_outputs);
-
 
     // TODO: Compute artificial viscosity once per timestep instead of each time
     // temperature system is assembled (as this might happen more than once per
@@ -3259,6 +1628,10 @@ namespace aspect
 #define INSTANTIATE(dim) \
   namespace internal { \
     namespace Assembly { \
+      namespace Scratch { \
+        template struct AdvectionSystem<dim>; \
+        template struct StokesSystem<dim>; \
+      } \
       template struct AssemblerLists<dim>::Properties; \
     } \
   } \
@@ -3278,10 +1651,6 @@ namespace aspect
   template void Simulator<dim>::copy_local_to_global_stokes_system ( \
                                                                      const internal::Assembly::CopyData::StokesSystem<dim> &data); \
   template void Simulator<dim>::assemble_stokes_system (); \
-  template void Simulator<dim>::get_artificial_viscosity (Vector<double> &viscosity_per_cell,  \
-                                                          const AdvectionField &advection_field) const; \
-  template void Simulator<dim>::get_artificial_viscosity (Vector<float> &viscosity_per_cell,  \
-                                                          const AdvectionField &advection_field) const; \
   template void Simulator<dim>::build_advection_preconditioner (const AdvectionField &, \
                                                                 std_cxx11::shared_ptr<aspect::LinearAlgebra::PreconditionILU> &preconditioner); \
   template void Simulator<dim>::local_assemble_advection_system ( \
