+++ tests/blankenbach-approximation.prm
@@ -1,0 +1,113 @@
+# Listing of Parameters
+# ---------------------
+# This case is identical to case 1a from the Blankenbach Benchmark
+# (Blankenbach 1989). Cases b and c can be run by changing the 
+# Viscosity to 1e-5 and 1e-6, respectively, which leads to the
+# corresponding change of the Rayleigh number from 1e4 to 1e5 and
+# 1e6. This test essentially resembles the convection-box cookbook
+# and reproduces its results very closely (a slight difference results
+# from the different refinement levels).
+# The relevant output is the vrms velocity and the heat flux through
+# the top boundary, which in this setup is identical to the Nu number. 
+
+set End time                               = 0.002 # 1.0 for full benchmark
+set Use years in output instead of seconds = false
+
+
+subsection Adiabatic conditions model
+  set Model name = function
+
+  subsection Function
+    set Function constants  = 
+    set Function expression = 0; 0
+    set Variable names      = depth
+  end
+
+end
+
+subsection Boundary temperature model
+  set Model name = box
+  subsection Box
+    set Bottom temperature = 1
+    set Left temperature   = 0
+    set Right temperature  = 0
+    set Top temperature    = 0
+  end
+end
+
+subsection Geometry model
+  set Model name = box 
+
+
+  subsection Box
+    set X extent = 1
+    set Y extent = 1
+  end
+end
+
+
+subsection Gravity model
+  set Model name = vertical
+  subsection Vertical
+    set Magnitude = 1e4 
+  end
+end
+
+
+subsection Initial conditions
+  set Model name = function
+  subsection Function
+    set Function constants  = z1=0.102367, z2=0.897633, pi=3.14159265359
+    set Function expression = if(z<z1,0.5+0.5*(z1-z)/z1,if(z>z2,0.5*(1-z)/(1-z2),0.5)) + 0.1 * cos(x*pi) * sin(z*pi)
+    set Variable names      = x,z
+  end
+end
+
+
+subsection Material model
+  set Model name = simple 
+
+  subsection Simple model
+    set Reference density             = 1
+    set Reference specific heat       = 1
+    set Reference temperature         = 0
+    set Thermal conductivity          = 1
+    set Thermal expansion coefficient = 1
+    set Viscosity                     = 1
+  end
+end
+
+
+subsection Mesh refinement
+  set Initial adaptive refinement        = 0
+  set Initial global refinement          = 5
+  set Time steps between mesh refinement = 0
+end
+
+subsection Heating model
+  set List of model names = #adiabatic heating, shear heating#, constant heating
+end
+
+subsection Model settings
+  set Include adiabatic heating               = false
+  set Include shear heating                   = false 
+  set Prescribed velocity boundary indicators = 
+  set Tangential velocity boundary indicators = left, right, bottom, top
+  set Zero velocity boundary indicators       = 
+  set Fixed temperature boundary indicators   = bottom, top
+end
+
+subsection Formulation
+  set Formulation = BA#custom
+  set Mass density approximation        = implicit adiabatic
+  set Temperature density approximation = adiabatic
+  set Buoyancy density approximation    = full#adiabatic pressure
+end
+
+subsection Postprocess
+  set List of postprocessors = visualization, velocity statistics, temperature statistics, heat flux statistics, heating statistics
+subsection Visualization
+    set Time between graphical output = 1e-10
+    set List of output variables      = material properties, adiabat, thermal conductivity, heating, artificial viscosity
+  end
+end
