+++ tests/tangurnis-mass-expl.cc
@@ -1,0 +1,1 @@
+#include "tangurnis-mass-impl.cc"
