+++ tests/tangurnis-mass-expl/statistics
@@ -1,0 +1,19 @@
+# 1: Time step number
+# 2: Time (seconds)
+# 3: Number of mesh cells
+# 4: Number of Stokes degrees of freedom
+# 5: Number of temperature degrees of freedom
+# 6: Iterations for Stokes solver
+# 7: Velocity iterations in Stokes preconditioner
+# 8: Schur complement iterations in Stokes preconditioner
+# 9: Time step size (seconds)
+# 10: Visualization file name
+0 0.000000000000e+00 256 2467 1089 32 15  3 0.000000000000e+00                                        "" 
+0 0.000000000000e+00   0    0    0 24 25 25 0.000000000000e+00                                        "" 
+0 0.000000000000e+00   0    0    0 18 19 19 0.000000000000e+00                                        "" 
+0 0.000000000000e+00   0    0    0 14 15 15 0.000000000000e+00                                        "" 
+0 0.000000000000e+00   0    0    0  9 10 10 0.000000000000e+00                                        "" 
+0 0.000000000000e+00   0    0    0  5  6  6 0.000000000000e+00                                        "" 
+0 0.000000000000e+00   0    0    0  3  4  4 0.000000000000e+00                                        "" 
+1 1.000000000000e-04 256 2467 1089  0  0  0 1.000000000000e-04 output-tangurnis-mass-expl/solution-00000 
+0 0.000000000000e+00   0    0    0  0  0  0 9.036399896181e-01 output-tangurnis-mass-expl/solution-00001
