+++ tests/tangurnis-mass-impl/screen-output
@@ -1,0 +1,32 @@
+
+Loading shared library <./libtangurnis-mass-impl.so>
+
+Number of active cells: 256 (on 5 levels)
+Number of degrees of freedom: 3,556 (2,178+289+1,089)
+
+*** Timestep 0:  t=0 seconds
+   Rebuilding Stokes preconditioner...
+   Solving Stokes system... 30+2 iterations.
+      Nonlinear Stokes residual: 0.0223157
+   Solving Stokes system... 0+0 iterations.
+      Nonlinear Stokes residual: 1.24703e-09
+
+   Postprocessing:
+     writing:                  output.csv
+     Writing graphical output: output-tangurnis-mass-impl/solution-00000
+
+*** Timestep 1:  t=0.0001 seconds
+   Solving Stokes system... 0+0 iterations.
+      Nonlinear Stokes residual: 1.24703e-09
+
+   Postprocessing:
+     writing:                  output.csv
+     Writing graphical output: output-tangurnis-mass-impl/solution-00001
+
+Termination requested by criterion: end time
+
+
++---------------------------------------------+------------+------------+
++---------------------------------+-----------+------------+------------+
++---------------------------------+-----------+------------+------------+
+
