+++ tests/tangurnis-mass-impl/statistics
@@ -1,0 +1,13 @@
+# 1: Time step number
+# 2: Time (seconds)
+# 3: Number of mesh cells
+# 4: Number of Stokes degrees of freedom
+# 5: Number of temperature degrees of freedom
+# 6: Iterations for Stokes solver
+# 7: Velocity iterations in Stokes preconditioner
+# 8: Schur complement iterations in Stokes preconditioner
+# 9: Time step size (seconds)
+# 10: Visualization file name
+0 0.000000000000e+00 256 2467 1089 32 18 3 0.000000000000e+00                                        "" 
+1 1.000000000000e-04 256 2467 1089  0  0 0 1.000000000000e-04 output-tangurnis-mass-impl/solution-00000 
+0 0.000000000000e+00   0    0    0  0  0 0 9.036399933912e-01 output-tangurnis-mass-impl/solution-00001
